refactor common functions
